move: RenameCollectionModal을 CollectionPage 폴더로 이동
